# golang/internal/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "article_api",
        "contact": {},
        "license": {
            "name": "ryuki"
        },
        "version": "2.0"
    },
    "paths": {
        "/": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "helloWorld"
                ],
                "summary": "helloを返す",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Message"
                        }
                    }
                }
            }
        },
        "/articles": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "db"
                ],
                "summary": "全ての記事のデータを返す",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/server.Title"
                            }
                        }
                    }
                }
            },
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "新しい記事の投稿",
                "parameters": [
                    {
                        "description": "Article Json",
                        "name": "article_json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.ArticleAllSteps"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.Message"
                        }
                    }
                }
            }
        },
        "/articles/{article_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "urlパラメータで指定された番号の記事を出力",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "article_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ArticleAllSteps"
                        }
                    }
                }
            },
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "記事の編集",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "article_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Article Json",
                        "name": "article_json",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/server.ArticleAllSteps"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.Message"
                        }
                    }
                }
            }
        },
        "/articles/{article_id}/{step_id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "article"
                ],
                "summary": "urlパラメータで指定された記事のステップを出力",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "article_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Step ID",
                        "name": "step_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.ArticleAllSteps"
                        }
                    }
                }
            }
        },
        "/likes/{article_id}": {
            "put": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "like"
                ],
                "summary": "いいね数の追加",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "article_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.Message"
                        }
                    }
                }
            },
            "delete": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "like"
                ],
                "summary": "いいね数の削除",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Article ID",
                        "name": "article_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/server.Message"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.Message"
                        }
                    }
                }
            }
        },
        "/swagger": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "helloWorld"
                ],
                "summary": "/swagger/index.html#/にアクセスするとswaggerを返す",
                "responses": {
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/server.Message"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "server.ArticleAllSteps": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "likes": {
                    "type": "integer"
                },
                "stepCount": {
                    "type": "integer"
                },
                "steps": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.Step"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "server.Code": {
            "type": "object",
            "properties": {
                "codeContent": {
                    "type": "string"
                },
                "codeFileName": {
                    "type": "string"
                }
            }
        },
        "server.Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
        "server.Step": {
            "type": "object",
            "properties": {
                "codes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/server.Code"
                    }
                },
                "content": {
                    "type": "string"
                }
            }
        },
        "server.Title": {
            "type": "object",
            "properties": {
                "articleId": {
                    "type": "string"
                },
                "author": {
                    "type": "string"
                },
                "likes": {
                    "type": "integer"
                },
                "stepCount": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        }
    }
}